Index theorems and definitions by default
This commit enables the theorem/rule/definition indexing by default,
and changes how it works (a little bit):
* By default, indexes are maintaind in the current workspace and $HOLDIR
  under a path .hol-vscode/. The user may add additional dependencies to
  a workspace by placing a file called dependencies.json in .hol-vscode,
  containing a JSON array of directory paths.
* If there is no existing index database for the current workspace and
  its dependencies then it is created.

Aside from this, the HOLIDE class is refactored, and documentation in
comments are added here and there.

The providers appear to be working most of the time, but fully qualified
identifiers (such as fooTheory.my_theorem) confuse the plugin a bit.